refactor: decompose oversized dashboard pages and fix Tidal polling leak
Fix a resource leak where Tidal auth polling intervals were never
cleaned up on component unmount. Move polling into a dedicated useEffect
keyed on tidalLoginPolling state with proper cleanup for both the
interval and the 10-minute timeout.

Replace 7 console.error swallows in events page handlers with
setActionError() calls so users see actionable error messages instead
of silent failures.

Extract RequestModal (207 lines) from display/page.tsx (1179→982 lines)
into its own component with self-contained state, search, submission,
and inactivity timeout logic.

Extract EventErrorCard component for the duplicated 410/404/fallback
error pattern used across authenticated pages.
